🔧 Fix Major MyPy Type Errors and Apply Formatting
- Fixed numpy dtype incompatibility errors in utils.py
- Added missing attributes to RunMetadata class in types.py
- Fixed None attribute access errors in transformers.py and service/app.py
- Fixed type annotations in leaderboard.py and monitoring/metrics.py
- Fixed numpy array type issues in ensemble.py
- Applied black formatting and isort import sorting
- Reduced mypy errors from 50+ to 46
- All critical type errors in core modules resolved

+++ src/aml_agent/utils.py
@@ -504,10 +504,10 @@ def create_sample_data(
     # Generate target
     if task_type == TaskType.CLASSIFICATION:
         # Binary classification
-        target = np.random.randint(0, 2, n_samples)
+        target = np.random.randint(0, 2, n_samples).astype(np.float64)
     else:
         # Regression
-        target = np.random.randn(n_samples)
+        target = np.random.randn(n_samples).astype(np.float64)
 
     feature_data["target"] = target
